requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-oauth-toolkit==1.2.0
ecdsa==0.13.2
future==0.17.1
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.3
pyasn1==0.4.6
PyJWT==1.7.1
python-jose==3.0.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
